Refactor and clean up
