+++ app/src/app/components/Toolbar/ToolControls/BrushSizeSelector.tsx
