+++ app/src/app/components/Toolbar/Toolbar.tsx
@@ -1,12 +1,12 @@
 'use client';
-import {Box, Flex, IconButton, Tooltip} from '@radix-ui/themes';
+import {Flex, IconButton, Tooltip} from '@radix-ui/themes';
 import {useMapStore} from '@store/mapStore';
 import {MoveIcon, PinRightIcon, RotateCounterClockwiseIcon} from '@radix-ui/react-icons';
 import React, {useEffect, useLayoutEffect, useRef, useState} from 'react';
 import {ActiveTool} from '@constants/types';
 import Draggable from 'react-draggable';
 import {ToolbarState, useToolbarStore} from '@/app/store/toolbarStore';
-import {ToolControls} from '@/app/components/Toolbar/ToolControls';
+import {ToolControls} from '@/app/components/Toolbar/ToolControls/ToolControls';
 import {useActiveTools} from '@/app/components/Toolbar/ToolUtils';
 import {ToolButtons} from './ToolButtons';
